% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munging.R
\name{compute_days}
\alias{compute_days}
\title{Compute a table of days from a table of actions.}
\usage{
compute_days(actions)
}
\arguments{
\item{actions}{Table of actions.}
}
\description{
Compute a table of days from a table of actions.
}

