% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munging.R
\name{compute_pages}
\alias{compute_pages}
\title{Compute a table of pages from a table of actions.}
\usage{
compute_pages(actions, base_url)
}
\arguments{
\item{actions}{Table of actions.}

\item{base_url}{A string like "host.com" with no trailing slash.}
}
\description{
Compute a table of pages from a table of actions.
}

