% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munging.R
\name{compute_sources}
\alias{compute_sources}
\title{Summarise visits by the source of traffic.}
\usage{
compute_sources(visits)
}
\arguments{
\item{visits}{Table of visits.}
}
\description{
Summarise visits by the source of traffic.
}

