% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munging.R
\name{compute_visitors}
\alias{compute_visitors}
\title{Compute a table of visitors from a table of actions.}
\usage{
compute_visitors(actions)
}
\arguments{
\item{actions}{Table of actions.}
}
\description{
Compute a table of visitors from a table of actions.
}

