% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\name{db_from_env}
\alias{db_from_env}
\title{Read environment variables MYSQL_(DBNAME|HOST|PORT|USER|PASSWORD)
    and connect to corresponding database.}
\usage{
db_from_env()
}
\description{
Read environment variables MYSQL_(DBNAME|HOST|PORT|USER|PASSWORD)
    and connect to corresponding database.
}

