% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_browser_resolutions}
\alias{graph_browser_resolutions}
\title{Graph distribution of browser resolutions.}
\usage{
graph_browser_resolutions(visits)
}
\arguments{
\item{visits}{Table of visits.}
}
\description{
Graph distribution of browser resolutions.
}

