% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_site_structure}
\alias{graph_site_structure}
\title{Graph structure of site based on visitor actions.}
\usage{
graph_site_structure(actions)
}
\arguments{
\item{actions}{Table of actions.}
}
\description{
Graph structure of site based on visitor actions.
}

