% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_visitors_vs_date}
\alias{graph_visitors_vs_date}
\title{Graph number of visitors over time.}
\usage{
graph_visitors_vs_date(days, point_size = 0.5)
}
\arguments{
\item{days}{Table of days.}

\item{point_size}{Size of the points plotted for each day.}
}
\description{
Graph number of visitors over time.
}

