% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{src_mysql}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{src_mysql}}}
}}
\keyword{internal}

