% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munging.R
\name{compute_pages}
\alias{compute_pages}
\title{Compute a table of pages from a table of actions.}
\usage{
compute_pages(actions)
}
\arguments{
\item{actions}{Table of actions.}
}
\description{
Compute a table of pages from a table of actions.
}

