% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_visitors_vs_date}
\alias{graph_visitors_vs_date}
\title{Graph number of visitors over time.}
\usage{
graph_visitors_vs_date(days)
}
\arguments{
\item{days}{Table of days.}
}
\description{
Graph number of visitors over time.
}

